% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CyclicVariationalCausalDiscovery}
\alias{CyclicVariationalCausalDiscovery}
\title{CyclicVariationalCausalDiscovery}
\description{
This class provides a variational inference algorithm for cyclic structrual equation models
}
\section{Fields}{

\describe{
\item{\code{new}}{Constructor}
}}

