\name{Rcpp_CyclicVariationalCausalDiscovery-class}
\Rdversion{1.1}
\docType{class}
\alias{Rcpp_CyclicVariationalCausalDiscovery-class}

\title{Class \code{"Rcpp_CyclicVariationalCausalDiscovery"}}
\description{
This class provides a variational inference algorithm for cyclic structrual equation models
}
\section{Extends}{
Class \code{"\linkS4class{C++Object}"}, directly.

All reference classes extend and inherit methods from \code{"\linkS4class{envRefClass}"}.

}

\author{
person(
    "Zhuofan",
    "Wang",
    ,
    "zf.wang@ruc.edu.cn",
    role = c("aut", "cre"))
}
%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
showClass("Rcpp_CyclicVariationalCausalDiscovery")
}
\keyword{classes}
\section{Methods}{
  \describe{
    \item{\code{initialize(...)}:}{ ~~ }
    \item{\code{finalize()}:}{ ~~ }
    \item{\code{run()}:}{ ~~ }
    \item{\code{init_run()}:}{ ~~ }
    \item{\code{get_pip()}:}{ ~~ }
    \item{\code{get_noiseParams()}:}{ ~~ }
    \item{\code{get_mu()}:}{ ~~ }
    \item{\code{get_A()}:}{ ~~ }
  }
}
